메뉴 수정 api 추가

+++ app/src/main/java/com/wafflestudio/siksha2/network/dto/LeaveReviewParam.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class LeaveReviewParam(
     @Json(name = "menu_id") val menuId: Long,
-    @Json(name = "score") val score: Double,
+    @Json(name = "score") val score: Long,
     @Json(name = "taste") val taste: String?,
     @Json(name = "price") val price: String?,
     @Json(name = "food_composition") val foodComposition: String?,
